requirements.txt
langchain-google-community==3.0.5
google-cloud-discoveryengine==0.16.0
google-genai==1.59.0